[ticket] 로그인 및 회원가입 페이지 리팩토링: useOAuth 훅 추가 및 리다이렉트 URI 환경 변수로 변경
